refactor: extract pure components
