+++ apps/cowswap-frontend/src/modules/hooksStore/pure/AppliedHookItem/styled.tsx
